Add optimized solution for day 1
Not much to optimize, mostly integer parsing and changing O(n^2) lookups with a pseudo-hashmap.
Using an actual hashmap seemed to be slower, especially for the sample cases, so I used a sorted vec and binary search for lookups.

Benchmarks indicated that part 1 (real input) went from 51us to 47us and part 2 from 75us to 50us, but the results were very noisy.