Add optimized solution for day 2
Mostly replacing a heap alloc with a stack alloc, and not checking the descending case if the ascending case already succeeds.
I tried using a single iteration that checked ascending, descending and jump all at once, but that turned out to be 20% slower.

Part 1 went from 103.92us to 46.275us (-55.5%) and part 2 went from 169.09us to 73.271us (-56.9%).