Add initial day 3 solution (position 483, 390)
Wow, a Regex challenge in AoC!? I was a bit surprised. Sure, you can manually parse this as well (probably optimal for runtime performance), but this is what I've been training for! <insert XKCD 208>

I just threw the input into regex101.com and threw together a regex as quickly as possible. Worked perfectly fine for part 1. The main difficulty was in looking up a usage example and quickly adding the dependency to the project.
For part 2 I was luckily already using `captures_iter`, which allowed me to add a "global" variable that I could flip during iteration of the captures, and I updated the regex to match any of the 3 instructions. Then I added a quick check on the third character to see which variant it matched and it's basically done.

      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  3   00:02:53   483      0   00:06:17   390      0