Optimize regex to not use unicode numbers
The Rust regex crate supports unicode by default, which means `\d` matches any unicode digit (which is more than just the ASCII ones). Changing it to specifically `0-9` gives a 30% and 22% performance improvement.
Switching from searching a &str to a &[u8] does not seem to give a measurable performance increase since we already have a &str, but it's done for completeness.

You can most definitely improve this by doing the parsing yourself instead of by using regex.