Add initial day 5 solution (positions 441, 358)
Whew, pretty happy with this. The code is pretty bad (got a chunk of code duplicated 3/4 times) but I got there pretty quickly! Still not enough for a top-100 finish, but there are plenty of days left in AoC.

I think this is one of the days that shows that Rust isn't great for quick-solving AoC, with the type checking, type annotations, lack of overloads (ie. `position` needs a lambda). Still, both parts being top500 is good!