Add initial optimization pass for day 5 part 1
All of this mostly comes down to reducing heap allocations and dynamically sized operations. It mostly comes from assuming that all numbers will be 2 digits long.

For rules, we can assume each line is 6 characters long (2 digits, a pipe, 2 digits, a newline) meaning we no longer have to seek to find any constant characters. We can get the length of the rules segment and then index into the slice. Parsing is as easy as getting the first and second digit and adding them together.
Rules are stored in an array of ArrayVecs. This is entirely stored on the heap and of fixed size. Each item in the array contains a list of the second part of the rule, where the index is the first part of the rule. 11 is subtracted from the indices because 11 is the first possible number.

Updates are no longer stored in a temporary vec (this was already redundant) and are now parsed in a similar way, by indexing into each line.
After parsing, they are checked against the rules by iterating over each number in the line and keeping a stack buffer of the numbers seen so far. This means we don't have to iterate over all the previous numbers in the line each time.

I've only optimized part 1 because part 2 is much of the same, and I have more stuff planned before I get to that.

Day 05 silver (sample)  time:   [196.46 ns 197.08 ns 197.83 ns]
                        change: [-84.896% -84.862% -84.827%] (p = 0.00 < 0.05)
                        Performance has improved.

Day 05 silver (real)    time:   [26.928 µs 26.975 µs 27.032 µs]
                        change: [-97.474% -97.467% -97.460%] (p = 0.00 < 0.05)
                        Performance has improved.