Improve performance of day 5 part 1 one final time
Last one, I promise. This is almost a full rewrite that aims to do all processing while it is parsing, using no intermediate buffers except the `rules` array with the bitmaps.
If a line is considered "bad", we enter a fast loop to search for the end of the line to skip any unneeded work.
Because we no longer have a list of numbers on each update line, we need to keep the index of the start of the line so we can work our way back to the middle number, as it's required for the sum.

There are still some tiny micro-optimizations to make (such as subtracting `b'00'` instead of `b'00' twice, removing bounds checks, etc) but I think this is a pretty good point. This is almost 200x faster than my initial implementation.

Day 05 silver (sample)  time:   [83.003 ns 83.060 ns 83.124 ns]
                        change: [-93.622% -93.613% -93.603%] (p = 0.00 < 0.05)
                        Performance has improved.

Day 05 silver (real)    time:   [5.6238 µs 5.6283 µs 5.6333 µs]
                        change: [-99.473% -99.472% -99.471%] (p = 0.00 < 0.05)
                        Performance has improved.