Add initial solution for day 9 (positions 705, 201)
I didn't expect my rank for part 2 to be so good! I guess this is one of those problems that LLM cheaters can't solve as easily?

I solved part 1 basically exactly the way it was described in the problem description so the code is quite simple but it's also fairly slow.
For part 2, I decided it was easier to store ranges (probably the intended solution) rather than individual nodes since then moving around entire files was easier. This made it harder to calculate the checksum at the end, so I just expanded the ranges into the full disk again.

      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  9   00:12:09   705      0   00:19:44   201      0